+++ cli_app/src/handlers/mod.rs
@@ -1,0 +1,4 @@
+pub mod autogen;
+pub mod seeder;
+pub mod uploader;
+pub mod validate;
